--- routes/api/signIn/index.js
@@ -1,7 +1,0 @@
-const router = require("express").Router();
-
-const { signInUser } = require("./controller");
-
-router.route("/").post(signInUser);
-
-module.exports = router;
